frontend/notifications: even more refactoring and tweaking queries
